Check it out, yo!
1.  Added file specification delimiters (All File Types, Include,
Exclude) to filter image files for processing.  The textbox for which
extensions to take into consideration is disabled if the user has chosen
to process All File Types.

2.  Added functionality to filter images to process by searching for a
string that is contained in their filename.

3.  Made a check box to give the user the option if they would like to
copy the original image over to the newly created directory (per Dr.
Piraino)

4.  Made browsing directories to search for Macro Files and R Scripts
(added them to UI)

5.  Created a reusable validation routine for user input.  I also
implemented two specific cases:
-if the user types in an invalid file extension
-if the user selects an invalid Macro file