Changed a lot with running macros
Wrote a class the extends the runnable interface, so that macro files
can be executed on a new thread.
The main menu creates a new thread to run the macro, when the macro is
done, that thread tells the main class to create a new thread for the
next macro.

Also somehow some of my regular expressions got deleted, so I put those
back in the generalizer