Allows user to see changes to the file when running the macro
The file is placed in a folder so when we are finished, we can place all
scripts in the folder creating a plugin with submenus. By using the view
step plugin, the user is greeted with a popup that tells them to press
ok to continue. This stops the macro and allows them to view any changes
the macro performed