Some new functionalities
1. created user_paths.txt file located in Fiji's plugins directory.
This file contains the paths for a user's most recently used
directories.  The fields for directories will auto-populate upon opening
so long as the user has previously used the application to process
images.

2. New file menu option - Change R Path.  If the user has data for a
path to RScript.exe (in user_paths.exe), the system will not prompt them
upon opening to navigate to the file.  However, if the user wishes to
change to path to RScript.exe, I've created a file menu option to
re-prompt the user.

3. Abstracted some of my code (e.g., writing to log file, creating and
showing an error dialog, etc.) into seperate methods in order to improve
our code's ease of readability.

4. many small bug fixes and code touch ups