Added automatic pressings of the enter key
IN ORDER TO RUN THE PIPEPLINE, YOU MUST HAVE Press_Enter.py in the
Medical_Image DIRECTORY.

Some functions in imagej bring up a dialog box that requires the user to
press ok. The pipeline now will automatically press enter if a new
process is not called within two seconds.