Changed the download procedure
The reason that the one image was having downloading issues was do to an
issue in the java ssl code. SSL Certificates with a length of 1024 or
greater just error out. As fiji uses its own version of java, it does
not seem there is much we can do other than catch the exception. Also
the start button does not activate correctly with the url list