Added a menu option to create a basic r script
The menu options asks the user to select a .csv file. The file is read
in a a menu with two drop down boxes is create (one drop down for x
variable and one for y). The drop down boxes are populated with the
column names from the .csv file. The selected column names are passed
into the function generateBasicRScript. The function does nothing yet,
we must decide how we want to implement this